requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.11.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1759648874697/work/certifi
click==8.3.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1744743067588/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
exceptiongroup==1.3.0
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1759948064277/work
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1759187078489/work
fsspec==2025.10.0
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1756739567949/work
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1756321760188/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_kiwisolver_1762488741/work
lightning-utilities==0.15.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1759055168201/work
matplotlib==3.10.7
mdurl==0.1.2
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
munkres==1.1.4
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1731521053/work
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1759265530216/work
patsy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_patsy_1760998269/work
pillow==12.0.0
ply @ file:///home/conda/feedstock_root/build_artifacts/ply_1733239724146/work
protobuf==6.33.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1758436411/work
PyQt5==5.15.11
PyQt5_sip==12.17.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1758891829620/work
rich==14.2.0
safetensors==0.6.2
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1715869161825/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1739790642651/work/dist/scipy-1.15.2-cp310-cp310-linux_x86_64.whl#sha256=9e52bad6c3294d1a5b04a13632118ca2157130603c6c018c2d710162b223b27e
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1733730015268/work
shellingham==1.5.4
simsimd==6.5.3
sip @ file:///home/conda/feedstock_root/build_artifacts/sip_1759437834046/work
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sniffio==1.3.1
statsmodels @ file:///home/conda/feedstock_root/build_artifacts/statsmodels_1759296953068/work
stringzilla==4.2.3
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946051654/work
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
timm==1.0.9
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tomli_1760014679/work
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchmetrics==1.8.2
torchvision==0.19.1+cu121
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1762506714284/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
triton==3.0.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1762268864031/work
Werkzeug==3.1.3
